Changed how reviews are selected, fixed some bugs.
- Changed how reviews are selected for inclusion in the ReviewCloud.
Previously, the "most helpful" reviews were selected, but this could
lead to misleading ReviewClouds if a game had a rash of
positive/negative reviews. ReviewCloud for Steam now selects reviews
according to the overall sentiment for the game. If the game has 80%
positive reviews, then ReviewCloud for Steam will select 80% positive
reviews, and 20% negative reviews.
- Fixed issue #1: A single review should not be processed more than
once.
- Fixed issue #2: Line-breaks within reviews are now handled properly.
- I also re-converted spaces to tabs after the last commit changed them
back to spaces. D'oh!